Refactor world validation into focused modules
